[Refactor] NetworkManager객체 제거, Repository / Router의 역할 재정의
- Networkrouter가 Decoding하는 역할까지 제네릭하게 담당하도록 역할 재정의
- Repository가 NetworkRouter를 직접 호출 (특정 타입 지정)
