+++ BoxOffice/Sources/Application/DIContainer/DIContainer.swift
@@ -15,8 +15,7 @@ final class DIContainer {
     
     func makeBoxOfficeListController() -> BoxOfficeListController {
         let router = NetworkRouter()
-        let manager = NetworkManager(router: router)
-        let repository = BoxOfficeListRepository(manager: manager)
+        let repository = BoxOfficeListRepository(router: router)
         let usecase = BoxOfficeListUsecase(repository: repository)
         let viewModel = BoxOfficeListViewModel(usecase: usecase)
         let viewController = BoxOfficeListController(viewModel: viewModel)
