+++ NetworkRouterTests/NetworkRouterTests.swift
@@ -25,7 +25,7 @@ final class NetworkRouterTests: XCTestCase {
         sut = NetworkRouter(session: urlSession)
         
         // when
-        sut.request(endpoint) { result in
+        sut.request(endpoint) { (result: Result<Data, NetworkError>) in
             
             // then
             switch result {
@@ -38,30 +38,6 @@ final class NetworkRouterTests: XCTestCase {
                 }
                 
                 XCTAssertEqual(statusCode, urlSession.failureStatusCode)
-            }
-            expectation.fulfill()
-        }
-        
-        wait(for: [expectation], timeout: 1)
-    }
-    
-    func test_request메서드가_성공했을때_data가_비어있지않는지_확인한다() {
-        
-        // given
-        let expectation = XCTestExpectation()
-        let endpoint = MockEndpoint()
-        let urlSession = MockURLSession(isFailRequest: false)
-        sut = NetworkRouter(session: urlSession)
-        
-        // when
-        sut.request(endpoint) { result in
-            
-            // then
-            switch result {
-            case .success(let data):
-                XCTAssertTrue(data.isEmpty == false)
-            case .failure:
-                XCTFail("Mock URLSession이 성공하는 세션이기에 Fail이면 안됨")
             }
             expectation.fulfill()
         }
